Fixes #29445 - restrict STI deletion to admin
When obsolete STI records deletion was introduced, we allowed even
non-admin users to delete that. This patch makes it possible only for
admins to delete records from the database. Non-admin users only see a
short message explaining what happened, that they should contact
administrator and request ID that can help administrator to figure out
what happened.

It was also necessary to add explicit require_login call, otherwise
User.current was nil in case the stack failed before user was loaded.
This also made the page looks much nicer, becase the menu can be
properly rendered.

+++ app/controllers/application_controller.rb
@@ -122,9 +122,10 @@ class ApplicationController < ActionController::Base
   end
 
   def sti_clean_up(e)
+    require_login rescue false
     @unknown_class_name = e.message.match(/subclass: '(.*)'\./)[1]
     @parent_class = e.message.match(/overwrite (.*)\.inheritance_column/)[1].constantize
-    if params[:confirm_data_deletion] == 'yes'
+    if params[:confirm_data_deletion] == 'yes' && User.current&.admin?
       begin
         @parent_class.where(@parent_class.inheritance_column => @unknown_class_name).delete_all
       rescue ActiveRecord::InvalidForeignKey => e
